Update autoported files

+++ itextsharp.kernel/itextsharp/kernel/xmp/impl/Base64.cs
@@ -43,13 +43,13 @@ namespace iTextSharp.Kernel.Xmp.Impl
 	public class Base64
 	{
 		/// <summary>marker for invalid bytes</summary>
-		private const byte INVALID = (byte)-1;
+		private const byte INVALID = -1;
 
 		/// <summary>marker for accepted whitespace bytes</summary>
-		private const byte WHITESPACE = (byte)-2;
+		private const byte WHITESPACE = -2;
 
 		/// <summary>marker for an equal symbol</summary>
-		private const byte EQUAL = (byte)-3;
+		private const byte EQUAL = -3;
 
 		private static byte[] base64 = new byte[] { (byte)'A', (byte)'B', (byte)'C', (byte
 			)'D', (byte)'E', (byte)'F', (byte)'G', (byte)'H', (byte)'I', (byte)'J', (byte)'K'
